AI_Voice_Recognition_Transcribe/requirements_transformers.txt
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
requests>=2.31.0
python-dotenv>=1.0.0
soundfile>=0.12.0